Renamed pcode128 test module to tests
